started camera class
camera class integrated

move keyboard/mouse input to separate class

tried to add texture

texture atlas coords figured out

cube stuff figured out

left side fixed

cube working with texture atlas

put texture in separate class

mesh class

backface culling

render multiple blocks

gitignore

+++ Clonecraft/headers/glitter.hpp
@@ -18,7 +18,7 @@
 #include <stb_image.h>
 
 // Define Some Constants
-const int mWidth = 1280;
-const int mHeight = 800;
+const float mWidth = 1280.0f;
+const float mHeight = 800.0f;
 
 #endif //~ Glitter Header
